some other edit

+++ resources/js/components/IndexComponent.vue
@@ -23,7 +23,7 @@
                 <td>{{ post.title }}</td>
                 <td>{{ post.body }}</td>
                 <td><router-link :to="{name: 'edit', params: { id: post.id }}" class="btn btn-primary">Edit</router-link></td>
-                <td><button class="btn btn-danger">Delete</button></td>
+                <td><button class="btn btn-danger" @click.prevent="deletePost(post.id)">Delete</button></td>
             </tr>
             </tbody>
         </table>
@@ -40,16 +40,24 @@ export default {
     created() {
         let uri = 'http://127.0.0.1:8000/api/posts';
         this.axios.get(uri).then(response => {
+            console.log(response);
             this.posts = response.data.data;
         });
     },
     methods: {
         deletePost(id)
         {
-            let uri = `http://127.0.0.1:8000/api/post/delete/${id}`;
-            this.axios.delete(uri).then(response => {
-                this.posts.splice(this.posts.indexOf(id), 1);
-            });
+            let r = confirm("Do you want to delete record?");
+            if(r === true) {
+                let uri = `http://127.0.0.1:8000/api/post/delete/${id}`;
+                this.axios.delete(uri).then(response => {
+                    console.log(this.posts);
+                    // this.posts.splice(this.posts.indexOf(id), 1);
+                    this.posts.filter(post => {
+                        return post.id !== id;
+                    });
+                });
+            }
         }
     }
 }
